requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==1.6.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
billiard==3.6.4.0
boto3==1.34.84
botocore==1.34.84
celery==5.2.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.39.1
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dj-database-url==0.5.0
Django==5.0.6
django-environ==0.8.1
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.15.1
dropbox==11.36.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isodate==0.6.1
jmespath==1.0.1
kombu==5.3.6
msgpack==1.0.8
msrest==0.7.1
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.3.0
ply==3.11
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.8.0
pycparser==2.22
PyMySQL==1.1.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2021.3
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.2
stone==3.3.3
toml==0.10.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==5.3.0
zope.interface==6.4.post2
pydub == 0.25.1
django-celery-beat
flower
channels_redis